requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.30.0
cffi==2.0.0
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
ecdsa==0.19.2
fastapi==0.115.12
greenlet==3.4.0
h11==0.16.0
httptools==0.7.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.3
pycparser==3.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.2.2
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.40
starlette==0.46.2
typing_extensions==4.15.0
uvicorn==0.34.0
watchfiles==1.1.1
websockets==16.0
